pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "realistic-al"
dynamic = ["version"]
requires-python= ">=3.8,<3.9"
dependencies = [
    "batchgenerators==0.24",
    "GitPython==3.1.27",
    "hydra-core==1.2.0",
    "lightning_bolts==0.5.0",
    "loguru==0.6.0",
    "matplotlib==3.5.1",
    "numpy==1.22.3",
    "omegaconf==2.2.2",
    "pandas==1.4.3",
    "Pillow==9.2.0",
    "pytest==7.1.2",
    "pytorch_lightning==1.6.4",
    "PyYAML==6.0",
    "rich==12.5.1",
    "scikit_learn==1.1.1",
    "scipy==1.7.3",
    "seaborn==0.11.2",
    "six==1.16.0",
    "toma==1.1.0",
    "torch==1.12.0",
    "torchmetrics==0.9.2",
    "torchvision==0.13.0",
    "tqdm==4.64.0"
]

[project.optional-dependencies]
dev = [
  "black",
  "debugpy",
  "ipython",
  "mypy",
  "pre-commit",
  "pylint",
  "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)